refactor: add menugroup package
